fastapi/requirements.txt
databricks-sql-connector==4.0.0
databricks-sdk==0.45.0
fastapi==0.115.11
posit-sdk==0.9.0
